requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
cryptography==37.0.4
distlib==0.3.5
dj-database-url==1.0.0
Django==3.2.7
django-crispy-forms==1.13.0
django-extensions==3.2.0
django-on-heroku==1.1.2
filelock==3.8.0
Flask==2.2.1
gunicorn==20.1.0
h11==0.13.0
idna==3.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
outcome==1.2.0
Pillow==8.3.1
pipenv==2022.8.19
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pyTelegramBotAPI==4.2.2
python-decouple==3.6
python-dotenv==0.20.0
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.4
requests==2.26.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
tabulate==0.8.9
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.6
virtualenv==20.16.3
virtualenv-clone==0.5.7
webdriver-manager==3.8.3
Werkzeug==2.2.1
whitenoise==6.2.0
wikipedia==1.4.0
wsproto==1.1.0
zipp==3.8.1
